refactor(lsp_server): organize commands into modules
